Updated prior code to utilize the new dataset
Optimized Joey's "main" code to include RandomSearchCV and a confusion matrix. Applied my updated version to Vu's dataset (where the original train dataset was split to a new "PEMS_train" and new "PEMS_test" to include only unique data). This was done to address the original test dataset containing duplicate data from the train data. With this model, applied to the non-duplicated dataset, the model resulted in an F1-score of 1.0.